requirements.txt
python-pptx==0.6.23
requests>=2.28.0
python-dotenv>=1.0.0
openai>=1.0.0
schedule>=1.2.0
streamlit>=1.28.0
